fix average reward in summary

+++ explorer/entity.go
@@ -203,7 +203,7 @@ type BlockTransactions struct {
 }
 
 type RewardSummary struct {
-	AverageReward                     int     `json:"averageReward"`
+	AverageReward                     float64 `json:"averageReward"`
 	AverageInput                      float64 `json:"averageInput"`
 	MedianInput                       float64 `json:"medianInput"`
 	AveragePoSInput                   float64 `json:"averagePoSInput"`
